add shared folder to organize common components and refactor components
