+++ src/components/Genre/index.tsx
@@ -1,0 +1,52 @@
+import React from 'react'
+
+import { SubmenuSidebar } from './styles'
+
+export function Genre() {
+  return (
+    <SubmenuSidebar>
+      <h4>Gêneros</h4>
+      <ul>
+        <li>
+          <a href="/#">Ação</a>
+        </li>
+        <li>
+          <a href="/#">Aventura</a>
+        </li>
+        <li>
+          <a href="/#">Animação</a>
+        </li>
+        <li>
+          <a href="/#">Comédia</a>
+        </li>
+        <li>
+          <a href="/#">Crime</a>
+        </li>
+        <li>
+          <a href="/#">Documentário</a>
+        </li>
+        <li>
+          <a href="/#">Drama</a>
+        </li>
+        <li>
+          <a href="/#">Família</a>
+        </li>
+        <li>
+          <a href="/#">Fantasia</a>
+        </li>
+        <li>
+          <a href="/#">Terror</a>
+        </li>
+        <li>
+          <a href="/#">Mistério</a>
+        </li>
+        <li>
+          <a href="/#">Romance</a>
+        </li>
+        <li>
+          <a href="/#">Ficção científica</a>
+        </li>
+      </ul>
+    </SubmenuSidebar>
+  )
+}
