+++ src/components/Preview/types.ts
@@ -1,5 +1,5 @@
 import { DiscoverResult } from '../../store/modules/discover/types'
 
-export interface CardProps {
-  movie: DiscoverResult
+export interface PreviewProps {
+  preview: DiscoverResult
 }
