+++ src/components/shared/Card/index.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+import { CardProps } from './types'
+import { Wrapper } from './styles'
+
+export function Card(props: CardProps) {
+  const { children } = props
+
+  return <Wrapper>{children}</Wrapper>
+}
