+++ src/components/shared/Card/types.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react'
+
+export interface CardProps {
+  children: ReactNode
+}
