+++ src/components/shared/Sidebar/types.ts
@@ -1,0 +1,9 @@
+import { ReactNode } from 'react'
+
+export interface SidebarProps {
+  children: ReactNode
+}
+
+export interface SidebarStylesProps {
+  isOpen: boolean
+}
